Move SqlNormalizer to package ...api.db.normalizer
