+++ instrumentation/cassandra/cassandra-4.0/src/main/java/io/opentelemetry/javaagent/instrumentation/cassandra/v4_0/CassandraDatabaseClientTracer.java
@@ -11,8 +11,8 @@ import com.datastax.oss.driver.api.core.cql.ExecutionInfo;
 import com.datastax.oss.driver.api.core.metadata.Node;
 import io.opentelemetry.instrumentation.api.tracer.DatabaseClientTracer;
 import io.opentelemetry.instrumentation.api.tracer.utils.NetPeerUtils;
+import io.opentelemetry.javaagent.instrumentation.api.db.CassandraQueryNormalizer;
 import io.opentelemetry.javaagent.instrumentation.api.db.DbSystem;
-import io.opentelemetry.javaagent.instrumentation.api.db.cassandra.CassandraQueryNormalizer;
 import io.opentelemetry.trace.Span;
 import java.net.InetSocketAddress;
 import java.util.Optional;
