+++ javaagent-api/javaagent-api.gradle
@@ -14,13 +14,13 @@ javacc {
   configs {
     template {
       inputFile = file('src/main/javacc/SqlNormalizer.jj')
-      packageName = 'io.opentelemetry.javaagent.instrumentation.api.db.sql.normalizer'
+      packageName = 'io.opentelemetry.javaagent.instrumentation.api.db.normalizer'
     }
   }
 }
 
 tasks.withType(Checkstyle).configureEach {
-  exclude '**/db/sql/normalizer/*.java'
+  exclude '**/db/normalizer/*.java'
 }
 
 dependencies {
