+++ javaagent-api/src/main/javacc/SqlNormalizer.jj
@@ -11,7 +11,7 @@ options {
 
 PARSER_BEGIN(SqlNormalizer)
 
-package io.opentelemetry.javaagent.instrumentation.api.db.sql.normalizer;
+package io.opentelemetry.javaagent.instrumentation.api.db.normalizer;
 
 import java.io.StringReader;
